# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: featured
          in: query
          required: false
          schema:
            type: boolean
        - name: limit
          in: query
          required: false
          schema:
            type: number
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Invalid product
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product
      parameters:
        - $ref: "#/components/parameters/ProductId"
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Product not found
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - $ref: "#/components/parameters/ProductId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Product not found
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - $ref: "#/components/parameters/ProductId"
      responses:
        "204":
          description: Product deleted
        "404":
          description: Product not found
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders for a customer
      parameters:
        - name: email
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Order list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid order
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order
      parameters:
        - $ref: "#/components/parameters/OrderId"
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get the demo customer profile
      responses:
        "200":
          description: Customer profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
    patch:
      operationId: updateProfile
      tags: [profile]
      summary: Update the customer profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
...
components:
  parameters:
    ProductId:
      name: id
      in: path
      required: true
      schema:
        type: number
        minimum: 1
    OrderId:
      name: id
      in: path
      required: true
      schema:
        type: number
        minimum: 1
...
  schemas:
...
    Product:
      type: object
      required: [id, name, slug, description, category, price, stock, imageUrl, featured, rating, reviewCount]
      properties:
        id: { type: number }
        name: { type: string }
        slug: { type: string }
        description: { type: string }
        category: { type: string }
        price: { type: number }
        compareAtPrice: { type: ["number", "null"] }
        stock: { type: number }
        imageUrl: { type: string }
        featured: { type: boolean }
        rating: { type: number }
        reviewCount: { type: number }
        createdAt: { type: string, format: date-time }
    ProductInput:
      type: object
      required: [name, description, category, price, stock, imageUrl]
      properties:
        name: { type: string, minLength: 2 }
        description: { type: string, minLength: 2 }
        category: { type: string, minLength: 2 }
        price: { type: number, minimum: 0 }
        compareAtPrice: { type: ["number", "null"], minimum: 0 }
        stock: { type: number, minimum: 0 }
        imageUrl: { type: string }
        featured: { type: boolean, default: false }
    ProductUpdate:
      allOf:
        - $ref: "#/components/schemas/ProductInput"
      description: Product fields to update
...
    ShippingOption:
      type: object
      required: [id, carrier, service, price, deliveryDays, description]
      properties:
        id: { type: string }
        carrier: { type: string }
        service: { type: string }
        price: { type: number }
        deliveryDays: { type: number }
        description: { type: string }
    OrderItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId: { type: number, minimum: 1 }
        quantity: { type: number, minimum: 1 }
    OrderItem:
      type: object
      required: [productId, productName, quantity, unitPrice, total]
      properties:
        productId: { type: number }
        productName: { type: string }
        quantity: { type: number }
        unitPrice: { type: number }
        total: { type: number }
    Order:
      type: object
      required: [id, customerName, customerEmail, status, subtotal, shipping, total, shippingOption, items, createdAt]
      properties:
        id: { type: number }
        customerName: { type: string }
        customerEmail: { type: string }
        status: { type: string }
        subtotal: { type: number }
        shipping: { type: number }
        total: { type: number }
        shippingOption: { $ref: "#/components/schemas/ShippingOption" }
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt: { type: string, format: date-time }
    OrderInput:
      type: object
      required: [customerName, customerEmail, postalCode, shippingOptionId, items]
      properties:
        customerName: { type: string, minLength: 2 }
        customerEmail: { type: string }
        postalCode: { type: string, minLength: 8 }
        shippingOptionId: { type: string }
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/OrderItemInput"
    Profile:
      type: object
      required: [id, name, email, phone, postalCode, city, state]
      properties:
        id: { type: number }
        name: { type: string }
        email: { type: string }
        phone: { type: string }
        postalCode: { type: string }
        city: { type: string }
        state: { type: string }
    ProfileUpdate:
      type: object
      required: [name, email, phone, postalCode, city, state]
      properties:
        name: { type: string, minLength: 2 }
        email: { type: string }
        phone: { type: string }
        postalCode: { type: string }
        city: { type: string }
        state: { type: string }
...